graspe_py/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rtb"
version = "0.1.0"
description = "Projeto Python usando Robotics Toolbox do Peter Corke"
authors = [
    { name="João Pedro Baltieca Garcia", email="joao_garciajp@usp.br" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "numpy>=1.25.0",
    "scipy>=1.15.3",
    "matplotlib>=3.10.6",
    "sympy>=1.12.0",
    "roboticstoolbox-python>=1.1.1",
    "spatialmath-python>=1.1.14",
    "pandas>=2.0.0",
    "pyyaml>=6.0.0"
]

[tool.setuptools]
package-dir = {"" = "src"} 
packages = ["graspe_py"]

[project.optional-dependencies]
dev = [
    "jupyterlab",
    "ipython",
    "pre-commit"
]
